HW_03_04_final/requirements.txt
alabaster==1.0.0
arrow==1.3.0
asgiref==3.8.1
asttokens==2.4.0
babel==2.16.0
binaryornot==0.4.4
bleach==6.2.0
bokeh==3.6.0
cachetools==5.5.0
certifi==2024.8.30
cfgv==3.4.0
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
colorcet==3.1.0
comm==0.1.4
ConfigUpdater==3.2
contourpy==1.3.0
cookiecutter==2.6.0
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
distlib==0.3.9
Django==4.2.16
docutils==0.21.2
exceptiongroup==1.1.2
executing==1.2.0
filelock==3.16.1
fonttools==4.54.1
holoviews==1.19.1
hvplot==0.11.1
identify==2.6.1
idna==3.10
imagesize==1.4.1
importlib_metadata==6.8.0
ipykernel==6.25.1
ipython==8.16.1
jedi==0.19.0
Jinja2==3.1.4
jupyter_client==8.3.0
jupyter_core==5.3.1
kiwisolver==1.4.7
linkify-it-py==2.0.3
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.2
matplotlib-inline==0.1.6
mdit-py-plugins==0.4.2
mdurl==0.1.2
myst-parser==4.0.0
nest_asyncio==1.5.6
nodeenv==1.9.1
numpy==2.1.2
packaging==24.1
pandas==2.2.3
panel==1.5.3
param==2.1.1
parso==0.8.3
pickleshare==0.7.5
Pillow==11.0.0
platformdirs==4.3.6
pluggy==1.5.0
pre_commit==4.0.1
prompt_toolkit==3.0.39
psutil==5.9.5
pure_eval==0.2.2
Pygments
pyparsing==3.2.0
pyproject-api==1.8.0
PyScaffold==4.6
pyscaffoldext-cookiecutter==0.1
pyscaffoldext-custom-extension==0.6.3
pyscaffoldext-django==0.2
pyscaffoldext-dsproject==0.7.2
pyscaffoldext-markdown==0.5
pyscaffoldext-travis==0.3
python-dateutil==2.8.2
python-slugify==8.0.4
pytz==2024.2
pyviz_comms==3.0.3
PyYAML==6.0.2
pyzmq==25.1.1
requests==2.32.3
rich==13.9.3
seaborn==0.13.2
setuptools==75.1.0
setuptools-scm==8.1.0
six==1.16.0
snowballstemmer==2.2.0
Sphinx==8.1.3
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
sqlparse==0.5.1
stack_data==0.6.2
text-unidecode==1.3
tomlkit==0.13.2
tornado==6.3.3
tox==4.23.2
tqdm==4.66.6
traitlets==5.11.2
types-python-dateutil==2.9.0.20241003
typing_extensions==4.8.0
tzdata==2024.2
uc-micro-py==1.0.3
urllib3==2.2.3
virtualenv==20.27.1
wcwidth==0.2.6
webencodings==0.5.1
wheel==0.44.0
xyzservices==2024.9.0
zipp==3.17.0
opencv-python
pyqt5